Add and style section 4
NOTE: The spacing in this page is becoming a mess, some things are
padded from the sides, some are just sitting in the center at fixed
object widths. It works, sort of, at half-monitor width, but it's
neither consistent nor pretty when resized significantly. This is in
part because I misunderstood the suggestion of going section-by-section
as meaning that each section should be directly a part in the body tag.
It could be done with a flexbox around everything, but it's not
necessary right now.